refine uart driver
Add hardware FIFO support

+++ Libraries/utilities/imu/imu.c
@@ -15,15 +15,15 @@
 
 typedef struct 
 {
-    float ax;
-    float ay;
-    float az;
-    float gx;
-    float gy;
-    float gz;
-    float mx;
-    float my;
-    float mz;
+    double ax;
+    double ay;
+    double az;
+    double gx;
+    double gy;
+    double gz;
+    double mx;
+    double my;
+    double mz;
 }imu_float_data_t;
 
 
@@ -100,43 +100,43 @@ static uint32_t imu_sliding_filter(imu_raw_data_t raw_data, imu_raw_data_t * fil
 //!< format data into float value
 static uint32_t imu_format_data(imu_raw_data_t * raw_data, imu_float_data_t * float_data)
 {
-    float_data->ax = (float)raw_data->ax;
-    float_data->ay = (float)raw_data->ay;
-    float_data->az = (float)raw_data->az;
+    float_data->ax = (double)raw_data->ax;
+    float_data->ay = (double)raw_data->ay;
+    float_data->az = (double)raw_data->az;
     
-    float_data->gx = (float)raw_data->gx;
-    float_data->gy = (float)raw_data->gy;
-    float_data->gz = (float)raw_data->gz;
+    float_data->gx = (double)raw_data->gx;
+    float_data->gy = (double)raw_data->gy;
+    float_data->gz = (double)raw_data->gz;
     
-    float_data->mx = (float)raw_data->mx;
-    float_data->my = (float)raw_data->my;
-    float_data->mz = (float)raw_data->mz;
+    float_data->mx = (double)raw_data->mx;
+    float_data->my = (double)raw_data->my;
+    float_data->mz = (double)raw_data->mz;
     return 0;
 }
 
 //!< the mx my mz order are related to PCB layout!!
-static void updateAHRS(float gx,float gy,float gz,float ax,float ay,float az,float mx,float mz,float my, imu_float_euler_angle_t * angle)
+static void updateAHRS(double gx,double gy,double gz,double ax,double ay,double az,double mx,double mz,double my, imu_float_euler_angle_t * angle)
 {
-    float norm = 0;
-    float hx = 0, hy = 0, hz = 0, bx = 0, bz = 0;			
-    float vx = 0, vy = 0, vz = 0, wx = 0, wy = 0, wz = 0;		
-    float ex = 0, ey = 0, ez = 0;
-    static float q0 = 1; 
-    static float q1 = 0; 
-    static float q2 = 0;
-    static float q3 = 0;
-    static float exInt = 0, eyInt = 0, ezInt = 0;
+    double norm = 0;
+    double hx = 0, hy = 0, hz = 0, bx = 0, bz = 0;			
+    double vx = 0, vy = 0, vz = 0, wx = 0, wy = 0, wz = 0;		
+    double ex = 0, ey = 0, ez = 0;
+    static double q0 = 1; 
+    static double q1 = 0; 
+    static double q2 = 0;
+    static double q3 = 0;
+    static double exInt = 0, eyInt = 0, ezInt = 0;
 
-    float q0q0 = q0*q0;
-    float q0q1 = q0*q1;
-    float q0q2 = q0*q2;
-    float q0q3 = q0*q3;
-    float q1q1 = q1*q1;
-    float q1q2 = q1*q2;
-    float q1q3 = q1*q3;
-    float q2q2 = q2*q2;
-    float q2q3 = q2*q3;
-    float q3q3 = q3*q3;
+    double q0q0 = q0*q0;
+    double q0q1 = q0*q1;
+    double q0q2 = q0*q2;
+    double q0q3 = q0*q3;
+    double q1q1 = q1*q1;
+    double q1q2 = q1*q2;
+    double q1q3 = q1*q3;
+    double q2q2 = q2*q2;
+    double q2q3 = q2*q3;
+    double q3q3 = q3*q3;
 
     if(ax*ay*az==0)
     {
@@ -196,6 +196,7 @@ static void updateAHRS(float gx,float gy,float gz,float ax,float ay,float az,flo
     q1 = q1 / norm;
     q2 = q2 / norm;
     q3 = q3 / norm;
+    
     /* output data */
     angle->imu_yaw = atan2(2 * q1 * q2 + 2 * q0 * q3, -2 * q2*q2 - 2 * q3* q3 + 1)* 57.3; 
     angle->imu_pitch = asin(-2 * q1 * q3 + 2 * q0* q2)* 57.3;																			// pitcho???
